reco_content/requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.2
djangorestframework==3.11.0
djangorestframework-firebase==0.3.0
firebase-admin==2.18.0
gast==0.3.3
gcloud==0.17.0
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.2
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.28.1
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
h5py==2.10.0
httplib2==0.17.3
idna==2.9
importlib-metadata==1.6.0
joblib==0.14.1
jws==0.1.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.2
msgpack==1.0.0
numpy==1.18.4
oauth2client==3.0.0
opencv-python==4.2.0.32
pandas==1.0.3
Pillow==5.1.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
tb-nightly==1.14.0a20190301
tensorflow==2.0.0a0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.1.0
